Separates timeseries modules and creates interface
