+++ pyfilter/timeseries/__init__.py
@@ -1,4 +1,4 @@
-from .base import AffineModel, AffineObservations
+from .affine import AffineModel, AffineObservations
 from .eulermaruyma import EulerMaruyma, OrnsteinUhlenbeck
 from .model import StateSpaceModel
 from .parameter import Parameter
